add sync-installation api
