+++ app/api/github/webhook/route.ts
@@ -2,7 +2,7 @@ import { NextRequest, NextResponse } from 'next/server';
 import { getRequestContext } from '@cloudflare/next-on-pages';
 import { getPrismaClient } from '@/lib/db';
 import { Webhooks } from '@octokit/webhooks';
-import { getInstallationAccessToken } from '@/lib/github';
+import { getInstallationAccessToken, findByrdocsArchiveFork, upsertInstallation } from '@/lib/github';
 
 export const runtime = 'edge';
 
@@ -52,119 +52,6 @@ async function verifyWebhookSignature(request: NextRequest, webhookSecret: strin
   } catch (error) {
     return { verified: false, error: 'Invalid signature' };
   }
-}
-
-// Helper function to check if a repository is a byrdocs-archive fork by fetching parent info from GitHub API
-async function isByrdocsArchiveFork(repoFullName: string, githubToken?: string): Promise<boolean> {
-  try {
-    const response = await fetch(`https://api.github.com/repos/${repoFullName}`, {
-      headers: {
-        'Accept': 'application/vnd.github.v3+json',
-        ...(githubToken && { 'Authorization': `Bearer ${githubToken}` })
-      },
-    });
-
-    if (!response.ok) {
-      console.warn(`Failed to fetch repo details for ${repoFullName}: ${response.status}`);
-      return false;
-    }
-
-    const repoData: any = await response.json();
-    
-    const isTargetFork = repoData.fork && repoData.parent?.full_name === `${process.env.NEXT_PUBLIC_ARCHIVE_REPO_OWNER}/${process.env.NEXT_PUBLIC_ARCHIVE_REPO_NAME}`;
-    
-    if (isTargetFork) {
-      console.log(`✓ Confirmed: ${repoFullName} is a fork of ${process.env.NEXT_PUBLIC_ARCHIVE_REPO_OWNER}/${process.env.NEXT_PUBLIC_ARCHIVE_REPO_NAME}`);
-    }
-    
-    return isTargetFork;
-  } catch (error) {
-    console.error(`Error checking if ${repoFullName} is ${process.env.NEXT_PUBLIC_ARCHIVE_REPO_NAME} fork:`, error);
-    return false;
-  }
-}
-
-// Helper function to find byrdocs-archive fork from repository list using heuristics
-async function findByrdocsArchiveFork(repositories?: Array<{ name: string; full_name: string; private: boolean }>, githubToken?: string): Promise<string | null> {
-  if (!repositories || repositories.length === 0) {
-    return null;
-  }
-
-  // Filter public repositories only (forks are always public)
-  const publicRepos = repositories.filter(repo => !repo.private);
-  
-  if (publicRepos.length === 0) {
-    return null;
-  }
-
-  console.log(`Checking ${publicRepos.length} public repositories for ${process.env.NEXT_PUBLIC_ARCHIVE_REPO_NAME} fork`);
-  
-  // Heuristic 1: Check repositories named 'byrdocs-archive' first
-  const namedRepos = publicRepos.filter(repo => repo.name === process.env.NEXT_PUBLIC_ARCHIVE_REPO_NAME);
-  const otherRepos = publicRepos.filter(repo => repo.name !== process.env.NEXT_PUBLIC_ARCHIVE_REPO_NAME);
-  
-  // First, check only repositories named 'byrdocs-archive'
-  if (namedRepos.length > 0) {
-    console.log(`Checking ${namedRepos.length} repositories named ${process.env.NEXT_PUBLIC_ARCHIVE_REPO_NAME} first`);
-    const namedRepoPromises = namedRepos.map(async (repo) => {
-      console.log(`Checking named repo: ${repo.full_name}`);
-      const isFork = await isByrdocsArchiveFork(repo.full_name, githubToken);
-      return isFork ? repo.name : null;
-    });
-
-    const namedResults = await Promise.all(namedRepoPromises);
-    const foundNamed = namedResults.find(result => result !== null);
-    
-    if (foundNamed) {
-      return foundNamed;
-    }
-  }
-  
-  // If no named repos found, check other repositories
-  if (otherRepos.length > 0) {
-    console.log(`No ${process.env.NEXT_PUBLIC_ARCHIVE_REPO_NAME} fork found in named repos, checking ${otherRepos.length} other repositories`);
-    const otherRepoPromises = otherRepos.map(async (repo) => {
-      console.log(`Checking other repo: ${repo.full_name}`);
-      const isFork = await isByrdocsArchiveFork(repo.full_name, githubToken);
-      return isFork ? repo.name : null;
-    });
-
-    const otherResults = await Promise.all(otherRepoPromises);
-    return otherResults.find(result => result !== null) || null;
-  }
-  
-  return null;
-}
-
-// Helper function to handle installation upsert
-async function upsertInstallation(
-  installationId: string,
-  accountLogin: string,
-  accountType: string,
-  repositoryName: string | null,
-  isSuspended: boolean = false
-) {
-  const prisma = getPrismaClient();
-  
-  return await prisma.gitHubInstallation.upsert({
-    where: {
-      installationId,
-    },
-    update: {
-      accountLogin,
-      accountType,
-      repositoryName,
-      isSuspended,
-      updatedAt: new Date(),
-    },
-    create: {
-      installationId,
-      accountLogin,
-      accountType,
-      repositoryName,
-      isSuspended,
-    },
-  });
 }
 
 // Main webhook handler
